+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html). 
-### [ita-challenges-frontend-2.1.0-RELEASE] - 2024-04-04
-* Adapt mobile navbar as figma design
+
+### [ita-challenges-frontend-2.2.0-RELEASE] - 2024-04-08
+* Challenge lists translates
 
 ### [ita-challenges-frontend-2.0.71-RELEASE] - 2024-04-02
 * Delete iteneraries service and implement the method on chalenges.service
@@ -13,6 +14,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### [ita-challenges-frontend-2.0.72-RELEASE] - 2024-04-03
 * Added responsiveness to login and register modals.
+
+### [ita-challenges-frontend-2.0.70-RELEASE] - 2024-03-21
+* Challenge list translates 
 
 ### [ita-challenges-frontend-2.0.69-RELEASE] - 2024-03-07
 * Adapt mobile-nav to Figma"